# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin + MongoDB构建的保险经纪公司管理平台",
        "title": "保险经纪管理系统API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户修改密码，需要提供当前密码和新密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "修改密码请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/YufungProject_internal_model.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "密码修改成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或token无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "当前密码错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "使用用户名和密码进行登录认证，返回JWT令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/YufungProject_internal_model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/YufungProject_internal_model.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "423": {
                        "description": "账户被锁定",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户退出登录，清除相关状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理"
                ],
                "summary": "用户登出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或token无效",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/YufungProject_internal_model.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "令牌刷新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/YufungProject_internal_model.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "刷新令牌无效或已过期",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "创建新用户账户，需要提供用户名、密码等基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证管理"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/YufungProject_internal_model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "注册成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/YufungProject_internal_model.UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "用户名或邮箱已存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/YufungProject_internal_model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "YufungProject_internal_model.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "description": "当前密码",
                    "type": "string"
                }
            }
        },
        "YufungProject_internal_model.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "YufungProject_internal_model.LoginResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "description": "令牌过期时间",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "刷新令牌",
                    "type": "string"
                },
                "token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "user": {
                    "description": "用户信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/YufungProject_internal_model.UserInfo"
                        }
                    ]
                }
            }
        },
        "YufungProject_internal_model.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "description": "刷新令牌",
                    "type": "string"
                }
            }
        },
        "YufungProject_internal_model.RegisterRequest": {
            "type": "object",
            "required": [
                "display_name",
                "password",
                "username"
            ],
            "properties": {
                "display_name": {
                    "description": "显示名称",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "email": {
                    "description": "邮箱地址",
                    "type": "string"
                },
                "password": {
                    "description": "密码",
                    "type": "string",
                    "minLength": 8
                },
                "phone": {
                    "description": "手机号码",
                    "type": "string"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                }
            }
        },
        "YufungProject_internal_model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                },
                "trace_id": {
                    "description": "追踪ID",
                    "type": "string"
                }
            }
        },
...
    }
}